Refactor and add functionalities to 'Notes' component
Refactored the existing 'Notes' component by removing the 'children' prop and including 'state', 'dispatch', and 'service' props. This was initiated to have more control over the state and actions on the component.

The changes also introduce functionalities to add, retrieve, and delete a note. 'Notes' is no longer a standalone component and has been integrated with 'Home' for better interaction between components.

Added two new components 'Note' and 'FormForNote' that are used in 'Notes' for displaying each note and adding new notes, respectively. This breaks down the code into smaller, manageable parts enhancing readability and maintainability.