Add update button and auto-focus on form
Included a button with refresh icon in the Notes component that calls the retrieveNotes function from the page component. This allows the user to update notes list. Removed unneeded console.log statements in page and Note components for cleaner code. Added focus to the text input of the FormForNote component after submitting, to improve user experience and streamline note submissions.