Refactor code to separate retrieveNotes function
This commit transfers the 'retrieveNotes' function from 'page.tsx' to a new location in 'retrieveNotes.ts' file in the 'utils' directory. 'retrieveNotes' is also updated in the import statements of each file that uses it. The aim of this refactor is to increase modularity and improve code organisation. It will also allow for potential future enhancements of the note receiving functionality.