Adjust top property for .form-button in globals.css
Incremented the top property of .form-button in globals.css. This adjustment was necessary to improve the position of the button, making it visually more appealing and user-friendly. This modification also contributes to the overall responsiveness of the user interface across various form layouts, enhancing user interaction especially in forms with numerous input fields.